pyproject.toml
[tool.poetry]
name = "minichat"
version = "1.0.1"
description = "Mini chatbot client based on AIML"
authors = ["DevAndromeda <devandromeda@snowflakedev.org>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/DevSnowflake/minichat"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [{ include = "minichat" }]

[tool.poetry.dependencies]
python = "^3.9"
python-aiml = "0.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
